Rename Python files to use snake case convention
Python files should be named using snake case convention.

- Rename a few Python files
- Adjust corresponding imports in `cook.py`
